Enable donation details in admin dash
