+++ src/app/donations/page.tsx
@@ -8,6 +8,8 @@ import '@/styles/globalStyles.css';
 import styles from '@/components/Browse.module.css';
 //Types
 import { Donation } from '@/models/donation';
+import { useState } from 'react';
+import DonationDetails from '@/components/DonationDetails';
 
 type DonationsProps = {
     donations: Donation[];
@@ -15,16 +17,24 @@ type DonationsProps = {
 
 const Donations = (props: DonationsProps) => {
     const { donations } = props;
+    const [showDetails, setShowDetails] = useState<string | null>(null);
+
     return (
         <>
-            <div className="page--header">
-                <h1>Donations</h1>
-            </div>
-            <ImageList className={styles['browse__grid']}>
-                {donations.map((donation) => (
-                    <DonationCard key={donation.id} donation={donation} />
-                ))}
-            </ImageList>
+            {showDetails ? (
+                <DonationDetails id={showDetails} />
+            ) : (
+                <>
+                    <div className="page--header">
+                        <h1>Donations</h1>
+                    </div>
+                    <ImageList className={styles['browse__grid']}>
+                        {donations.map((donation) => (
+                            <DonationCard key={donation.id} donation={donation} setShowDetails={setShowDetails} />
+                        ))}
+                    </ImageList>
+                </>
+            )}
         </>
     );
 };
